Introduce Test Coverage in Repo (#27)
This PR introduces MSTest in repo with emulator and cloud support. Two
sample tests are included as a potential template. Once this PR is in we
can slowly backfill testing gap.

------------------
Code has been structured in a way so as we can also boot emulator
standalone with included PS script and run the program against it
without having to use testing framework.